Made FuSM non-deterministic, Added Triangular membership, Added Irritated state

+++ FuSM/FuSMstate.h
@@ -24,10 +24,10 @@
 class FuSMstate  
 { 
 	int m_iStateID;				// the unique ID of this state
-	int m_iLowRange;			// range of int values to represent membership in this state
-	int m_iHighRange;
+	int m_iLowRange;			// Lower Limit for state
+	int m_iHighRange;			// Upper Limit for State
 	int m_iValueOfMembership;	// current amount of membership, for this fuzzy state
-	int m_iDegreeOfMembership;	// current percentage of truth, for this fuzzy state
+	float m_iDegreeOfMembership;	// current percentage of truth, for this fuzzy state
 
 public:
 	FuSMstate( int iStateID, int iLowRange, int iHighRange );
